Use case diagram for our website
The diagram depicts the choices a user can make in our website. They will be able to login with a username/password combination as a form of authentication credentials. From there, the user is taken to a home page where they can view, add, or delete videos presented to them.